CRUD advancements

+++ src/main/java/app/summarease/model/entities/chapter/dto/ChapterDto.java
@@ -1,12 +1,11 @@
 package app.summarease.model.entities.chapter.dto;
 
-public record ChapterDto(Integer id,
-
+public record ChapterDto(Long id,
                          String title,
                          String description,
                          boolean isNumbered,
                          String imageUrl,
-                         Integer parentDocumentId,
-                         Integer parentChapterId
-                         ) {
+                         Long parentDocumentId,
+                         Long parentChapterId
+) {
 }
